app rename sweep
This is a general sweep of whatever had the morphe-cli name.
