+++ .github/ISSUE_TEMPLATE/bug_report.yml
@@ -6,12 +6,12 @@ body:
   - type: markdown
     attributes:
       value: |
-        # Morphe CLI bug report
+        # Morphe Desktop bug report
 
         Before creating a new bug report, please keep the following in mind:
 
-        - **Do not submit a duplicate bug report**: Search for existing bug reports [here](https://github.com/MorpheApp/morphe-cli/issues?q=label%3A%22Bug+report%22).
-        - **Review the contribution guidelines**: Make sure your bug report adheres to it. You can find the guidelines [here](https://github.com/MorpheApp/morphe-cli/blob/main/CONTRIBUTING.md).
+        - **Do not submit a duplicate bug report**: Search for existing bug reports [here](https://github.com/MorpheApp/morphe-desktop/issues?q=label%3A%22Bug+report%22).
+        - **Review the contribution guidelines**: Make sure your bug report adheres to it. You can find the guidelines [here](https://github.com/MorpheApp/morphe-desktop/blob/main/CONTRIBUTING.md).
   - type: textarea
     attributes:
       label: Bug description
