+++ .github/ISSUE_TEMPLATE/feature_request.yml
@@ -6,12 +6,12 @@ body:
   - type: markdown
     attributes:
       value: |
-        # Morphe CLI feature request
+        # Morphe Desktop feature request
 
         Before creating a new feature request, please keep the following in mind:
 
-        - **Do not submit a duplicate feature request**: Search for existing feature requests [here](https://github.com/MorpheApp/morphe-cli/issues?q=label%3A%22Feature+request%22).
-        - **Review the contribution guidelines**: Make sure your feature request adheres to it. You can find the guidelines [here](https://github.com/MorpheApp/morphe-cli//blob/main/CONTRIBUTING.md).
+        - **Do not submit a duplicate feature request**: Search for existing feature requests [here](https://github.com/MorpheApp/morphe-desktop/issues?q=label%3A%22Feature+request%22).
+        - **Review the contribution guidelines**: Make sure your feature request adheres to it. You can find the guidelines [here](https://github.com/MorpheApp/morphe-desktop/blob/main/CONTRIBUTING.md).
   - type: textarea
     attributes:
       label: Feature description
