--- .run/CLI GUI.run.xml
@@ -1,9 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="CLI GUI" type="Application" factoryName="Application">
-    <option name="MAIN_CLASS_NAME" value="app.morphe.MorpheLauncherKt" />
-    <module name="morphe-cli.main" />
-    <method v="2">
-      <option name="Make" enabled="true" />
-    </method>
-  </configuration>
-</component>
