+++ CONTRIBUTING.md
@@ -1,32 +1,32 @@
 # 👋 Contribution guidelines
 
-This document describes how to contribute to Morphe CLI.
+This document describes how to contribute to Morphe Desktop.
 
 ## 📖 Resources to help you get started
 
-- The [documentation](/docs) explains how to use Morphe CLI
-- [Issues](https://github.com/MorpheApp/morphe-cli/issues) are where we keep track of bugs and feature requests.
+- The [documentation](/docs) explains how to use Morphe Desktop
+- [Issues](https://github.com/MorpheApp/morphe-desktop/issues) are where we keep track of bugs and feature requests.
 
 ## 🙏 Submitting a feature request
 
 Features can be requested by opening an issue using the
-[Feature request issue template](https://github.com/MorpheApp/morphe-cli/issues/new?assignees=&labels=Feature+request&projects=&template=feature_request.yml&title=feat%3A+).
+[Feature request issue template](https://github.com/MorpheApp/morphe-desktop/issues/new?assignees=&labels=Feature+request&projects=&template=feature_request.yml&title=feat%3A+).
 
 ## 🐞 Submitting a bug report
 
-If you encounter a bug while using Morphe CLI, open an issue using the
-[Bug report issue template](https://github.com/MorpheApp/morphe-cli/issues/new?assignees=&labels=Bug+report&projects=&template=bug_report.yml&title=bug%3A+).
+If you encounter a bug while using Morphe Desktop, open an issue using the
+[Bug report issue template](https://github.com/MorpheApp/morphe-desktop/issues/new?assignees=&labels=Bug+report&projects=&template=bug_report.yml&title=bug%3A+).
 
 ## 📝 How to contribute
 
 1. Before contributing, it is recommended to open an issue to discuss your change with the
-   maintainers of Morphe CLI. This will help you determine whether your change is acceptable
+   maintainers of Morphe Desktop. This will help you determine whether your change is acceptable
    and whether it is worth your time to implement it.
 2. Development happens on the `dev` branch. Fork the repository and create your branch from `dev`.
 3. Commit your changes.
 4. Submit a pull request to the `dev` branch of the repository and reference issues
    that your pull request closes in the description of your pull request.
 5. Our team will review your pull request and provide feedback. Once your pull request is approved,
-   it will be merged into the `dev` branch and will be included in the next release of Morphe CLI
+   it will be merged into the `dev` branch and will be included in the next release of Morphe Desktop
 
-❤️ Thank you for considering contributing to Morphe CLI.
+❤️ Thank you for considering contributing to Morphe Desktop.
