+++ settings.gradle.kts
@@ -11,7 +11,7 @@ plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
 }
 
-rootProject.name = "morphe-cli"
+rootProject.name = "morphe-desktop"
 
 // Include morphe-patcher and morphe-library as composite builds if they exist locally
 mapOf(
