+++ src/main/kotlin/app/morphe/gui/ui/screens/home/HomeScreen.kt
@@ -61,9 +61,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.ui.platform.LocalUriHandler
-import app.morphe.morphe_cli.generated.resources.Res
-import app.morphe.morphe_cli.generated.resources.morphe_dark
-import app.morphe.morphe_cli.generated.resources.morphe_light
+import app.morphe.morphe_desktop.generated.resources.Res
+import app.morphe.morphe_desktop.generated.resources.morphe_dark
+import app.morphe.morphe_desktop.generated.resources.morphe_light
 import app.morphe.gui.ui.theme.LocalMorpheCorners
 import app.morphe.gui.ui.theme.LocalMorpheDimens
 import app.morphe.gui.ui.theme.LocalMorpheFont
